refactor(bout-engine): decompose into validation + execution modules (RD-010)
- Extract bout-validation.ts: types, helpers, validateBoutRequest() (440 lines)
- Extract bout-execution.ts: executeBout(), settlement, error recovery (590 lines)
- Reduce bout-engine.ts to barrel file re-exporting both modules (27 lines)
- No external API changes: all consumers import from @/lib/bout-engine unchanged
- No test changes required: all 1419 tests pass, typecheck clean
